+++ terraform_compliance/steps/then/it_must_be_in.py
@@ -1,9 +1,13 @@
 from terraform_compliance.common.error_handling import Error
+from terraform_compliance.common.helper import Null
 
 def it_must_be_in(_step_obj, haystack, cover=False):
     match = _step_obj.context.match
 
     values = [resource['values'] for resource in _step_obj.context.stash]
+    if 'haystack' not in _step_obj.context.in_step_variables:
+        Error(_step_obj, 'Could not resolve haystack: {}'.format(haystack))       
+        return
     parsed_haystack = _step_obj.context.in_step_variables['haystack']
 
     if not cover and not match.issubset(values, parsed_haystack):
@@ -11,17 +15,3 @@ def it_must_be_in(_step_obj, haystack, cover=False):
 
     if cover and not match.issubset(parsed_haystack, values):
         Error(_step_obj, 'Selected resources does not cover {}.'.format(haystack))                                                                        
-
-'''
-inherited values?
-What if provider?
-
-a covers b
-subset(b, a)
-
-a in b
-subset(a, b)
-
-
-'''
-